feat: refactor lazy object template to prevent maximum call stack size exceeded error
